Fix issue with uptodate calls
Up-to-date checking used to short circuit calls, so if first item
in the `uptodate` list returned `False`, subsequent items were not
called. This introduced weird behaviors when using checkers with
side effect such as `result_dep`; in that case, outcome depended
on items order.

This patch removes the "short-circuiting" logic and forces every
up-to-date checker to be called.