Modify test_control.py to use a fixture to create sample tasks.
Task.init_options() will only initialize options once. The tests
were written so that each test reused the same task instances and
therefore the tests failed because the expected task option defaults
were not applied. The fix is to use a fixture to redefine the tasks
each time.

The only-once semantics of Task.init_options() dates back four years
and the commit indicates this was a performance enhancement.