tests: Make test_sqlite_import more robust and less demanding
test_sqlite_import requires that module sqlite3 has not been imported,
previously.  Dependening on the order in which tests are executed,
it may be the first to use SQLite or it may not.  That's unfortunate.
Examples of other tests using SQLite are all tests using the pdep_manager
fixture.  We could adjust ordering tests by means of pytest-dependency but
it won't be pretty.

Instead, Let's adjust test_sqlite_import to check for robust on-demand
import of SQLite, and refrain from pulluting the code base, just to
satisfy the demands of this single test.